Change structure of the integration test package
