--- integration/ce_test.go
@@ -1,9 +1,0 @@
-package integration
-
-import "testing"
-
-func TestCommunity(t *testing.T) {
-	if !isCE() {
-		t.Skip("skipping community edition tests")
-	}
-}
