+++ integration/operator_context_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/TykTechnologies/tyk-operator/api/model"
 	"github.com/TykTechnologies/tyk-operator/api/v1alpha1"
+	"github.com/TykTechnologies/tyk-operator/integration/common"
 	"github.com/matryer/is"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/e2e-framework/klient/k8s"
@@ -17,24 +18,18 @@ import (
 	"sigs.k8s.io/e2e-framework/pkg/features"
 )
 
-const (
-	testApiDef       = "test-http"
-	testOperatorCtx  = "mycontext"
-	gatewayLocalhost = "http://localhost:7000"
-)
-
 func TestOperatorContextCreate(t *testing.T) {
 	opCreate := features.New("Operator Context").
 		Setup(func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-			testNS := ctx.Value(ctxNSKey).(string)
+			testNS := ctx.Value(common.CtxNSKey).(string)
 			is := is.New(t)
 
 			// create operator context
-			opCtx, err := createTestOperatorContext(ctx, testNS, envConf)
+			opCtx, err := common.CreateTestOperatorContext(ctx, testNS, envConf)
 			is.NoErr(err) // failed to create operatorcontext
 
 			// create api definition
-			_, err = createTestAPIDef(ctx, testNS, func(apiDef *v1alpha1.ApiDefinition) {
+			_, err = common.CreateTestAPIDef(ctx, testNS, func(apiDef *v1alpha1.ApiDefinition) {
 				apiDef.Spec.Context = &model.Target{
 					Name:      opCtx.Name,
 					Namespace: opCtx.Namespace,
@@ -43,7 +38,7 @@ func TestOperatorContextCreate(t *testing.T) {
 			is.NoErr(err) // failed to create apiDefinition
 
 			// create api definition with empty namespace for contextRef
-			_, err = createTestAPIDef(ctx, testNS, func(apiDef *v1alpha1.ApiDefinition) {
+			_, err = common.CreateTestAPIDef(ctx, testNS, func(apiDef *v1alpha1.ApiDefinition) {
 				apiDef.Name = "empty-ns"
 				apiDef.Spec.Context = &model.Target{Name: opCtx.Name}
 				apiDef.Spec.Proxy.ListenPath = "/empty-ns"
@@ -55,11 +50,11 @@ func TestOperatorContextCreate(t *testing.T) {
 		}).
 		Assess("context status is updated",
 			func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-				testNS := ctx.Value(ctxNSKey).(string) //nolint:errcheck
+				testNS := ctx.Value(common.CtxNSKey).(string) //nolint:errcheck
 				client := envConf.Client()
 				is := is.New(t)
 
-				opCtx := v1alpha1.OperatorContext{ObjectMeta: metav1.ObjectMeta{Name: testOperatorCtx, Namespace: testNS}}
+				opCtx := v1alpha1.OperatorContext{ObjectMeta: metav1.ObjectMeta{Name: common.TestOperatorCtx, Namespace: testNS}}
 
 				err := wait.For(conditions.New(client.Resources()).ResourceMatch(&opCtx, func(object k8s.Object) bool {
 					operatCtx := object.(*v1alpha1.OperatorContext) //nolint:errcheck
@@ -71,12 +66,12 @@ func TestOperatorContextCreate(t *testing.T) {
 					}
 
 					if operatCtx.Status.LinkedApiDefinitions[0].Namespace != testNS ||
-						operatCtx.Status.LinkedApiDefinitions[0].Name != testApiDef {
+						operatCtx.Status.LinkedApiDefinitions[0].Name != common.TestApiDef {
 						return false
 					}
 
 					return true
-				}), wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+				}), wait.WithTimeout(common.DefaultWaitTimeout), wait.WithInterval(common.DefaultWaitInterval))
 				is.NoErr(err)
 
 				return ctx
@@ -85,7 +80,7 @@ func TestOperatorContextCreate(t *testing.T) {
 			is := is.New(t)
 
 			err := wait.For(func() (done bool, err error) {
-				resp, getErr := http.Get(fmt.Sprintf("%s/httpbin/get", gatewayLocalhost))
+				resp, getErr := http.Get(fmt.Sprintf("%s/httpbin/get", common.GatewayLocalhost))
 				if getErr != nil {
 					t.Log(getErr)
 					return false, nil
@@ -97,7 +92,7 @@ func TestOperatorContextCreate(t *testing.T) {
 				}
 
 				return true, nil
-			}, wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+			}, wait.WithTimeout(common.DefaultWaitTimeout), wait.WithInterval(common.DefaultWaitInterval))
 			is.NoErr(err)
 
 			return ctx
@@ -106,7 +101,7 @@ func TestOperatorContextCreate(t *testing.T) {
 			is := is.New(t)
 
 			err := wait.For(func() (done bool, err error) {
-				resp, getErr := http.Get(fmt.Sprintf("%s/empty-ns/get", gatewayLocalhost))
+				resp, getErr := http.Get(fmt.Sprintf("%s/empty-ns/get", common.GatewayLocalhost))
 				if getErr != nil {
 					t.Log(getErr)
 					return false, nil
@@ -118,7 +113,7 @@ func TestOperatorContextCreate(t *testing.T) {
 				}
 
 				return true, nil
-			}, wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+			}, wait.WithTimeout(common.DefaultWaitTimeout), wait.WithInterval(common.DefaultWaitInterval))
 
 			is.NoErr(err)
 
@@ -131,17 +126,17 @@ func TestOperatorContextCreate(t *testing.T) {
 func TestOperatorContextDelete(t *testing.T) {
 	delApiDef := features.New("Api Definition Delete").
 		Setup(func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-			testNS := ctx.Value(ctxNSKey).(string)
+			testNS := ctx.Value(common.CtxNSKey).(string)
 			is := is.New(t)
 			client := envConf.Client()
 
-			operatorCtx, err := createTestOperatorContext(ctx, testNS, envConf)
+			operatorCtx, err := common.CreateTestOperatorContext(ctx, testNS, envConf)
 			is.NoErr(err) // failed to create operatorcontext
 
-			ctx = context.WithValue(ctx, ctxOpCtxName, operatorCtx.Name)
+			ctx = context.WithValue(ctx, common.CtxOpCtxName, operatorCtx.Name)
 
 			// create api definition
-			def, err := createTestAPIDef(ctx, testNS, func(apiDef *v1alpha1.ApiDefinition) {
+			def, err := common.CreateTestAPIDef(ctx, testNS, func(apiDef *v1alpha1.ApiDefinition) {
 				apiDef.Spec.Context = &model.Target{
 					Name:      operatorCtx.Name,
 					Namespace: operatorCtx.Namespace,
@@ -149,7 +144,7 @@ func TestOperatorContextDelete(t *testing.T) {
 			}, envConf)
 			is.NoErr(err) // failed to create apiDefinition
 
-			ctx = context.WithValue(ctx, ctxApiName, def.Name)
+			ctx = context.WithValue(ctx, common.CtxApiName, def.Name)
 
 			err = wait.For(conditions.New(client.Resources()).ResourceMatch(operatorCtx, func(object k8s.Object) bool {
 				opCtx := object.(*v1alpha1.OperatorContext) //nolint:errcheck
@@ -160,15 +155,15 @@ func TestOperatorContextDelete(t *testing.T) {
 				}
 
 				return true
-			}), wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+			}), wait.WithTimeout(common.DefaultWaitTimeout), wait.WithInterval(common.DefaultWaitInterval))
 
 			is.NoErr(err)
 
 			return ctx
-		}).Assess("context ref should not get deleted while it is still been refered",
+		}).Assess("context ref should not get deleted while it is still been referred",
 		func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-			testNS := ctx.Value(ctxNSKey).(string)        //nolint:errcheck
-			opCtxName := ctx.Value(ctxOpCtxName).(string) //nolint:errcheck
+			testNS := ctx.Value(common.CtxNSKey).(string)        //nolint:errcheck
+			opCtxName := ctx.Value(common.CtxOpCtxName).(string) //nolint:errcheck
 
 			client := envConf.Client()
 			is := is.New(t)
@@ -184,9 +179,9 @@ func TestOperatorContextDelete(t *testing.T) {
 		}).
 		Assess("delete api def should delete operator context",
 			func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-				testNS := ctx.Value(ctxNSKey).(string)        //nolint:errcheck
-				opCtxName := ctx.Value(ctxOpCtxName).(string) //nolint:errcheck
-				apiDefName := ctx.Value(ctxApiName).(string)  //nolint:errcheck
+				testNS := ctx.Value(common.CtxNSKey).(string)        //nolint:errcheck
+				opCtxName := ctx.Value(common.CtxOpCtxName).(string) //nolint:errcheck
+				apiDefName := ctx.Value(common.CtxApiName).(string)  //nolint:errcheck
 
 				client := envConf.Client()
 				is := is.New(t)
@@ -198,7 +193,7 @@ func TestOperatorContextDelete(t *testing.T) {
 				is.NoErr(err)
 
 				err = wait.For(conditions.New(client.Resources()).ResourceDeleted(&opCtx),
-					wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+					wait.WithTimeout(common.DefaultWaitTimeout), wait.WithInterval(common.DefaultWaitInterval))
 				is.NoErr(err)
 
 				return ctx
@@ -207,17 +202,17 @@ func TestOperatorContextDelete(t *testing.T) {
 
 	updateApiDef := features.New("Remove context from Api Definition").
 		Setup(func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-			testNS := ctx.Value(ctxNSKey).(string)
+			testNS := ctx.Value(common.CtxNSKey).(string)
 			is := is.New(t)
 			client := envConf.Client()
 
-			operatorCtx, err := createTestOperatorContext(ctx, testNS, envConf)
+			operatorCtx, err := common.CreateTestOperatorContext(ctx, testNS, envConf)
 			is.NoErr(err) // failed to create operatorcontext
 
-			ctx = context.WithValue(ctx, ctxOpCtxName, operatorCtx.Name)
+			ctx = context.WithValue(ctx, common.CtxOpCtxName, operatorCtx.Name)
 
 			// create api definition
-			def, err := createTestAPIDef(ctx, testNS, func(apiDef *v1alpha1.ApiDefinition) {
+			def, err := common.CreateTestAPIDef(ctx, testNS, func(apiDef *v1alpha1.ApiDefinition) {
 				apiDef.Spec.Context = &model.Target{
 					Name:      operatorCtx.Name,
 					Namespace: operatorCtx.Namespace,
@@ -225,7 +220,7 @@ func TestOperatorContextDelete(t *testing.T) {
 			}, envConf)
 			is.NoErr(err) // failed to create apiDefinition
 
-			ctx = context.WithValue(ctx, ctxApiName, def.Name)
+			ctx = context.WithValue(ctx, common.CtxApiName, def.Name)
 
 			err = wait.For(conditions.New(client.Resources()).ResourceMatch(operatorCtx, func(object k8s.Object) bool {
 				opCtx := object.(*v1alpha1.OperatorContext) //nolint:errcheck
@@ -235,15 +230,15 @@ func TestOperatorContextDelete(t *testing.T) {
 				}
 
 				return true
-			}), wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+			}), wait.WithTimeout(common.DefaultWaitTimeout), wait.WithInterval(common.DefaultWaitInterval))
 
 			is.NoErr(err)
 
 			return ctx
 		}).Assess("context ref should not get deleted while it is still been refered",
 		func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-			testNS := ctx.Value(ctxNSKey).(string)        //nolint:errcheck
-			opCtxName := ctx.Value(ctxOpCtxName).(string) //nolint:errcheck
+			testNS := ctx.Value(common.CtxNSKey).(string)        //nolint:errcheck
+			opCtxName := ctx.Value(common.CtxOpCtxName).(string) //nolint:errcheck
 
 			client := envConf.Client()
 			is := is.New(t)
@@ -253,16 +248,16 @@ func TestOperatorContextDelete(t *testing.T) {
 			err := client.Resources().Delete(ctx, &opCtx)
 			is.NoErr(err)
 
-			err = client.Resources(testNS).Get(ctx, testOperatorCtx, testNS, &opCtx)
+			err = client.Resources(testNS).Get(ctx, common.TestOperatorCtx, testNS, &opCtx)
 			is.NoErr(err)
 
 			return ctx
 		}).
 		Assess("removing reference from apiDefinition should delete operator context",
 			func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-				testNS := ctx.Value(ctxNSKey).(string)        //nolint:errcheck
-				opCtxName := ctx.Value(ctxOpCtxName).(string) //nolint:errcheck
-				apiDefName := ctx.Value(ctxApiName).(string)  //nolint:errcheck
+				testNS := ctx.Value(common.CtxNSKey).(string)        //nolint:errcheck
+				opCtxName := ctx.Value(common.CtxOpCtxName).(string) //nolint:errcheck
+				apiDefName := ctx.Value(common.CtxApiName).(string)  //nolint:errcheck
 
 				client := envConf.Client()
 				is := is.New(t)
@@ -279,7 +274,7 @@ func TestOperatorContextDelete(t *testing.T) {
 				is.NoErr(err)
 
 				err = wait.For(conditions.New(client.Resources()).ResourceDeleted(&opCtx),
-					wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+					wait.WithTimeout(common.DefaultWaitTimeout), wait.WithInterval(common.DefaultWaitInterval))
 
 				is.NoErr(err)
