--- integration/pro_test.go
@@ -1,11 +1,0 @@
-package integration
-
-import (
-	"testing"
-)
-
-func TestPro(t *testing.T) {
-	if !isPro() {
-		t.Skip("Skipping pro tests")
-	}
-}
